refactor: update styling and structure of SecurityIntelligence component for improved clarity and responsiveness
