+++ src/components/landing/SecurityIntelligence.tsx
@@ -1,7 +1,5 @@
 "use client";
 import React from "react";
-import { Card, CardHeader } from "@heroui/react";
-import { Button } from "@heroui/react";
 
 // Define types for the fingerprint data
 type FingerprintItem = {
@@ -9,114 +7,82 @@ type FingerprintItem = {
   value: string;
 };
 
-type DeviceFingerprint = {
-  vpnStatus: string;
-  deviceType: string;
-  timestamp: string;
-  osModel: string;
-  ipAddress: string;
-  regionCity: string;
-  browserInfo: string;
-};
-
 function SecurityIntelligence() {
-  // Sample data - in a real app, this would come from an API or context
-  const fingerprintData: DeviceFingerprint = {
-    vpnStatus: "Not Detected",
-    deviceType: "Desktop",
-    timestamp: new Date().toLocaleString(),
-    osModel: "Win32",
-    ipAddress: "192.168.x.x (masked)",
-    regionCity: "Unknown",
-    browserInfo: "Mozilla/5.0 (Windows NT 10.0; Win64; x64...",
-  };
-
-  // Convert the data to a format easier to map through in the UI
+  // Table data exactly matching the screenshot
   const fingerprintItems: FingerprintItem[] = [
-    { label: "VPN Status", value: fingerprintData.vpnStatus },
-    { label: "Device Type", value: fingerprintData.deviceType },
-    { label: "Timestamp", value: fingerprintData.timestamp },
-    { label: "OS Model", value: fingerprintData.osModel },
-    { label: "IP Address", value: fingerprintData.ipAddress },
-    { label: "Region/City", value: fingerprintData.regionCity },
-    { label: "Browser Information", value: fingerprintData.browserInfo },
+    { label: "VPN Status", value: "Not Detected" },
+    { label: "Device Type", value: "Desktop" },
+    { label: "Timestamp", value: "5/9/2025, 7:39:09 AM" },
+    { label: "OS Model", value: "Win 32" },
+    { label: "IP address", value: "192.168.x.x(masked)" },
+    { label: "Region/City", value: "Unknown" },
+    {
+      label: "Browser Information",
+      value: "Mozilla/5.0( Windows NT 10.0: Win64: x64",
+    },
+    { label: "Device Trust Score", value: "92%" },
   ];
 
   return (
-    <div className="w-full h-auto bg-white">
-      <section className="w-full max-w-[1440px] mx-auto h-auto py-12">
-        <div className="h-auto w-full flex items-center flex-col py-2">
-          <div className="container w-full h-full">
-            {/* Header */}
-            <div className="w-full h-fit flex flex-col items-center justify-center gap-4">
-              <div className="w-fit h-fit font-Poppins text-slate-600 tracking-wider text-2xl xl:text-4xl font-semibold px-4 text-center">
-                Security Intelligence In Action
+    <div className="w-full h-auto bg-[#EFF6FF]">
+      <section className="w-full max-w-[1440px] mx-auto py-12 px-4">
+        {/* Header */}
+        <div className="text-center">
+          <h2 className="text-2xl md:text-3xl lg:text-4xl font-semibold text-slate-800 mb-2">
+            Security Intelligence in Action
+          </h2>
+          <div className="flex justify-center items-center mb-6">
+            <div className="w-28 h-1 bg-emerald-400 mx-auto"></div>
+          </div>
+        </div>
+
+        {/* Subheading */}
+        <div className="text-center mb-12">
+          <p className="text-base md:text-lg text-slate-600 max-w-3xl mx-auto">
+            Experience how Scrubbe collects and analyzes device and network data
+            to enhance security monitoring.
+          </p>
+        </div>
+
+        {/* Main Content - Card */}
+        <div className="max-w-4xl mx-auto">
+          <div className="bg-white rounded-xl shadow-sm overflow-hidden">
+            <div className="p-6">
+              <div className="mb-6">
+                <h3 className="text-lg font-bold text-slate-800">
+                  Your Device FingerPrint
+                </h3>
+                <p className="text-sm text-slate-600 mt-1">
+                  Below is a sample of the type of data Scrubbe can collect and
+                  analyse to identify Potentially Security Threats
+                </p>
               </div>
-              <div className="bg-emerald-500 h-1 w-16 rounded-full"></div>
-            </div>
 
-            {/* Subheading */}
-            <div className="w-full h-full flex items-center justify-center py-4">
-              <span className="h-fit xl:w-3/6 text-wrap text-center text-[1rem] md:text-[1.3rem]  text-muted-foreground px-2 xl:px-0 font-Raleway font-medium">
-                Experience how Scrubbe collects and analyzes device and network
-                data to enhance security monitoring.
-              </span>
-            </div>
-
-            {/* Main Content */}
-            <div className="w-full flex items-center justify-center mt-6">
-              <div className="w-11/12 h-fit">
-                <Card className="w-full">
-                  <CardHeader>
-                    <div className="w-full p-4">
-                      <div className="w-full h-fit space-y-2">
-                        <div className="w-fit h-fit font-Poppins text-slate-800 font-bold text-lg">
-                          Your Device Fingerprint
-                        </div>
-                        <div className="text-sm font-Poppins font-normal text-muted-foreground">
-                          Below is a sample of the type of data Scrubbe can
-                          collect and analyze to identify potential security
-                          threats:
-                        </div>
-                      </div>
-
-                      {/* 2-Column Layout for Data - Now dynamic */}
-                      <aside className="w-full mt-8 border border-blue-500 rounded-lg overflow-hidden">
-                        <table className="w-full border-collapse">
-                          <tbody>
-                            {fingerprintItems.map((item, index) => (
-                              <tr
-                                key={item.label}
-                                className="group hover:bg-slate-50 transition-colors"
-                              >
-                                <td className="py-4 px-6 border-b border-blue-500 border-r font-Poppins text-sm font-semibold text-slate-700 bg-slate-50 w-1/3">
-                                  {item.label}
-                                </td>
-                                <td
-                                  className={`py-4 px-6 font-Poppins text-md font-normal ${
-                                    index < fingerprintItems.length - 1
-                                      ? "border-b border-blue-500"
-                                      : ""
-                                  }`}
-                                >
-                                  {item.value}
-                                </td>
-                              </tr>
-                            ))}
-                          </tbody>
-                        </table>
-                      </aside>
-                    </div>
-                  </CardHeader>
-                </Card>
-                <div className="w-full flex justify-center mt-4">
-                  <Button
-                    className="bg-blue-600 hover:bg-blue-800 text-white font-semibold font-Inter text-md rounded-sm"
-                    size="lg"
-                  >
-                    Start free trial
-                  </Button>
-                </div>
+              {/* Data Table */}
+              <div className="border border-blue-100 rounded-lg overflow-hidden">
+                <table className="w-full border-collapse">
+                  <tbody>
+                    {fingerprintItems.map((item, index) => (
+                      <tr
+                        key={item.label}
+                        className="hover:bg-slate-50 transition-colors"
+                      >
+                        <td className="py-4 px-6 border-b border-blue-100 border-r font-medium text-sm text-slate-700 bg-slate-50 w-1/3">
+                          {item.label}
+                        </td>
+                        <td
+                          className={`py-4 px-6 text-md ${
+                            index < fingerprintItems.length - 1
+                              ? "border-b border-blue-100"
+                              : ""
+                          }`}
+                        >
+                          {item.value}
+                        </td>
+                      </tr>
+                    ))}
+                  </tbody>
+                </table>
               </div>
             </div>
           </div>
